map: do not check for bounds when linearly mapping values to ids
Files in /etc/iproute2 are linearly mapped in xtables_lmap_init().
Value ranges are limited to 0..255 range in iproute2 package.

However not all values are limited to this range. For example
network device group and realms stored as u32, so there is no
good reason on limiting these values to 0..255 range.

Currently there is no places in xtables where value in linear mapping
should be strictly in described range, so we could remove it safely.

In the future, any code which relies on value to be in given range
should check this by itself and do not depend on such check is
performed by the xtables_lmap_init().

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>