xtables: Support default action for the extension targets
Most of the targets implemented as an extension within kernel does not
explicitly determine fate of the packet making it to continue rule
traversal.

Sometimes we want to either accept or drop packet after being processed
by the target. One of the good examples for such extension targets are
MARK and TEE.

In some cases we might want to

  * accept packet after mark being set
  * drop packet after it's clone being sent to the gateway

Before this change this is only possible with adding new rule and very
likely new user-defined chain. That seems too complex. Instead introduce
new option --action that can be either "accept" or "drop" to explicitly
define fate of the packet processed by extension target.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>